test(shell): split the function-definition tests out of tests.rs
tests.rs reached 1565 lines and failed the LOC gate (ci.yml: ./scripts/loc-gate.sh,
limit 1500). 08c391b7 had trimmed the file below the gate; 3aae8096 pushed it back
over with the #1488/#1489 tests. Those move to function_def_tests.rs, following the
#[path] submodule pattern the file already uses for comment_strip/substitution/powershell.
